refactor(Olimpo): change models and add faccade
